CancerRAG/requirements-gpu.txt
langchain-ollama
langchain-huggingface
langchain-core
langchain-aws
langchain_community
faiss-gpu
beautifulsoup4
pandas
urllib3
matplotlib
tqdm
openpyxl
streamlit

langserve
uvicorn
fastapi 
sse_starlette